モジュール化対応 constants, logic, view 機能別にファイルを分割
