+++ public/index.html
@@ -19,7 +19,10 @@
       </form>
 
     </div>
-    <script src="js/app.js"></script>
+    <script type="module">
+      import { initPage } from './js/app.js';
+      initPage();
+  　</script>
   </div>
 </body>
