+++ public/js/constants.js
@@ -1,0 +1,7 @@
+const rot13Input = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz'.split('');
+const rot13Output = 'NOPQRSTUVWXYZABCDEFGHIJKLMnopqrstuvwxyzabcdefghijklm'.split('');
+
+export {
+  rot13Input,
+  rot13Output,
+};
